Create a split length field.
Change the logic of the length fields. Now the length field is split for a
chunk. Ten bits are now in the CIDTable. If the chunk is greater than 2 ^ 10
bytes, then another length field is created within the object. In addition, the
possibility of a memory allocation has been created that does not include a
length field.

The files:

*de/hhu/bsinfo/dxram/mem/SmallObjectHeapAnalyzer.java
*de/hhu/bsinfo/soh/SmallObjectHeapDumpAnalyzer.java
*de/hhu/bsinfo/soh/SmallObjectHeapIntegrityTest.java

have been deleted because they are no longer
usable due to the changes